feat: added indenization and premium automatic generation

+++ Classes/Apolice/Models/Apolice.cs
@@ -3,15 +3,15 @@ public class Apolice
   public int id_apolice { get; set; }
   public string data_inicio { get; set; }
   public string data_fim { get; set; }
-  public double premio { get; set; }
-  public double indenizacao { get; set; }
+  public decimal premio { get; set; }
+  public decimal indenizacao { get; set; }
   public int id_cobertura { get; set; }
   public int id_usuario { get; set; }
   public int id_cliente { get; set; }
   public int id_veiculo { get; set; }
   public string status { get; set; }
 
-  public Apolice(string startDate, string endDate, double premium, double indemnity, int idcobertura, int idusuario, int idcliente, int idveiculo, string status)
+  public Apolice(string startDate, string endDate, decimal premium, decimal indemnity, int idcobertura, int idusuario, int idcliente, int idveiculo, string status)
   {
 
     this.data_inicio = startDate;
